refactoring directory structure
