+++ backend/src/app.controller.ts
@@ -1,7 +1,7 @@
 import { Get, Controller, Render, Param, UseGuards, UseFilters, Session, Query, Res } from '@nestjs/common';
 import { AppService } from '../src/app.service';
-import { NFR_VARIANTS, NFR_VARIANTS_ARRAY, PAGE_TITLES, REPORT_TYPES } from 'utils/constants';
-import { SessionData } from 'utils/types';
+import { NFR_VARIANTS, NFR_VARIANTS_ARRAY, PAGE_TITLES, REPORT_TYPES } from 'src/constants';
+import { SessionData } from 'src/types';
 import { AuthenticationGuard } from './authentication/authentication.guard';
 import { AuthenticationFilter } from './authentication/authentication.filter';
 import { TTLSService } from './ttls/ttls.service';
@@ -11,7 +11,7 @@ import { firstValueFrom } from 'rxjs';
 import { Req } from '@nestjs/common/decorators/http/route-params.decorator';
 import { Request, Response } from 'express';
 import { ReportService } from './report/report.service';
-import { nfrInterestedParties } from 'utils/util';
+import { nfrInterestedParties } from 'src/util';
 
 //
 let requestUrl: string;
